requirements.txt
streamlit>=1.32
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.20.0
statsmodels>=0.14.0
openpyxl>=3.1.2
Flask>=3.0.0
# Add your framework here based on your model:
# scikit-learn>=1.3.0
tensorflow>=2.12.0
# torch>=2.0.0
# onnxruntime>=1.15.0
